rename ImageSavior to ImageSaver
